% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{CompileAllData}
\alias{CompileAllData}
\title{Just a wrapper to  combine all the data gathering functions}
\usage{
CompileAllData()
}
\description{
Just a wrapper to  combine all the data gathering functions
}
