% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verbs_filtered}
\alias{verbs_filtered}
\title{A sample data set of most frequent Russian verbs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 167 rows and 3 columns.}
\source{
The Araneum Russicum Maisu II corpus \url{http://ella.juls.savba.sk/aranea/index.html}
}
\usage{
verbs_filtered
}
\description{
The data set has been manually filtered to contain 100 only transitive non-auxiliary verbs (cf. column 'accept')
}
\keyword{datasets}
